+++ backend/models/prenotazione.js
@@ -34,18 +34,19 @@ prenotazioneSchema.pre('deleteOne', async function (next) {
     }
 });
 
-prenotazioneSchema.post('deleteOne', async function (next){
+prenotazioneSchema.post('deleteOne', async function (res, next){
     try {
         // Access the captured document
         const docToDelete = this._docToDelete;
 
-        console.log("PRE prenotazione");
         // Delete the reference to this booking for the user
         const output = await mongoose.model('utente').updateOne({ _id: docToDelete.utente }, { $pull: { prenotazioni: docToDelete._id } });
 
-        // Delete the reference to this booking for the report
-        const reportToDelete = docToDelete.report;
-        await mongoose.model('report').deleteOne({ _id: docToDelete.reportToDelete });
+        if(docToDelete.report){
+            // Delete the reference to this booking for the report
+            const reportToDelete = docToDelete.report;
+            await mongoose.model('report').deleteOne({ _id: docToDelete.reportToDelete });
+        }
 
         next(); // Call next to proceed to the next middleware
     } catch (error) {
